kadai 1

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+         $user =\Auth::user();
+         $tasks = $user->tasks()->paginate(10);
+         
+         $data = [
+             'user' => $user,
+             'tasks' => $tasks,
+             ];
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        return view('tasks.index',$data
+        );
+        }else {
+            return view('welcome');
+        }
         
     }
 
@@ -51,10 +61,10 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
 
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+       $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
 
         return redirect('/');
         //
@@ -66,16 +76,15 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+
     public function show($id)
     {
         $task = Task::find($id);
 
         return view('tasks.show', [
             'task' => $task,
-        ]);
-        //
+        ]);  
     }
-
     /**
      * Show the form for editing the specified resource.
      *
@@ -123,10 +132,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $micropost = \App\Task::find($id);
 
-        return redirect('/');
-        //
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        return redirect()->back();
     }
 }
